Improve accuracy of PIF EEPROM status command
Based on EEPROM test ROM run on real hardware.

+++ src/device/cart/eeprom.c
@@ -40,7 +40,7 @@ void format_eeprom(uint8_t* mem, size_t size)
 
 
 void init_eeprom(struct eeprom* eeprom,
-    uint32_t type,
+    uint16_t type,
     void* storage,
     const struct storage_backend_interface* istorage)
 {
